mega commit, bassoonoid is working but still very much WIP
- new task manager design with critical / normal classes. but it's not robust enough to handle very dense execution like we currently have so i plan to change it to some kind of static plan.
- fixes in settings multi items
- tests now work
- added circular array
- added performance monitoring menu app
- added individual subitem timings for musical state
- added NoInterrupts class
- added breath calibration setting item
- breath scaling now works better, also added curve
- added multi label setting item
- synth graph is now mono

+++ src/clarinoid/basic/Stopwatch.hpp
@@ -161,7 +161,7 @@ struct PeriodicTimer
   }
 
   int32_t BeatsToMicros(float b) const {
-    return b * mPeriodMicros;
+    return (int32_t)(b * mPeriodMicros);
   }
 };
